workers/constructors/SharedWorker/global-members.html

<!doctype html>
<title>members of SharedWorkerGlobalScope</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(function() {
  const worker = new SharedWorker('global-members.js');
  worker.port.onmessage = this.step_func_done(function(e) {
    assert_equals(e.data, '');
  });
}, 'Test if global members exist in a shared worker');
</script>
