w3c-reports/12092014/efl/runned-tests/workers/constructors/SharedWorker/global-members-expected.txt


FAIL members of SharedWorkerGlobalScope assert_equals: expected "" but got "applicationCache did not exist\n"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/workers/constructors/SharedWorker/global-members.html:22:18
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
